+++ keyboards/mb44/keymaps/default/keymap.c
@@ -19,8 +19,7 @@
 enum layer_names {
     _BASE,
     _LAYER1,
-    _LAYER2,
-    _LAYER3
+    _LAYER2
 
 };
 
@@ -40,34 +39,28 @@ qk_tap_dance_action_t tap_dance_actions[] = {
 
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-    [_BASE] = LAYOUT_DEFAULT(
-        KC_ESC,        KC_Q,          KC_W,    KC_E,                KC_R, KC_T, KC_Y,   KC_U, KC_I,    KC_O,    KC_P,    KC_BSPC,
-        CTL_T(KC_TAB), KC_A,          KC_S,    KC_D,                KC_F, KC_G, KC_H,   KC_J, KC_K,    KC_L,    KC_SCLN, KC_ENT,
-        KC_LSFT,       KC_Z,          KC_X,    KC_C,                KC_V, KC_B, KC_N,   KC_M, KC_COMM, KC_DOT,  KC_SLSH, RSFT_T(KC_UP),
-        KC_MPLY,       MO(_LAYER2),   KC_LALT, LT(_LAYER1, KC_SPC),             LT(_LAYER3, KC_SPC),                KC_LEFT, KC_RGHT, KC_DOWN
+    [_BASE] = LAYOUT_625U_SPACE(
+        KC_ESC,        KC_Q,    KC_W,    KC_E, KC_R, KC_T, KC_Y,          KC_U, KC_I,    KC_O,    KC_P,    KC_BSPC,
+        CTL_T(KC_TAB), KC_A,    KC_S,    KC_D, KC_F, KC_G, KC_H,          KC_J, KC_K,    KC_L,    KC_SCLN, KC_ENT,
+        KC_LSFT,       KC_Z,    KC_X,    KC_C, KC_V, KC_B, KC_N,          KC_M, KC_COMM, KC_DOT,  KC_SLSH, RSFT_T(KC_UP),
+        MO(_LAYER2),   KC_LGUI, KC_LALT,             LT(_LAYER1, KC_SPC),                KC_LEFT, KC_RGHT, KC_DOWN
     ),
 
-    [_LAYER1] = LAYOUT_DEFAULT(
-        KC_GRV,  KC_QUOT, XXXXXXX, KC_UP,   XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX,    KC_7, KC_8, KC_9,    KC_DEL,
-        KC_CAPS, XXXXXXX, KC_LEFT, KC_DOWN, KC_RGHT, XXXXXXX, XXXXXXX, TD(TD_M_D), KC_4, KC_5, KC_6,    KC_ENT,
-        KC_LSFT, KC_MINS, KC_EQL,  KC_LBRC, KC_RBRC, KC_BSLS, XXXXXXX, TD(TD_P_M), KC_1, KC_2, KC_3,    KC_RSFT,
-        XXXXXXX, KC_LGUI, KC_LALT, _______,                   XXXXXXX,                   KC_0, KC_PDOT, KC_PEQL
+    [_LAYER1] = LAYOUT_625U_SPACE(
+        KC_GRV,  KC_QUOT, _______, KC_UP,   _______, _______, _______, _______,    KC_7, KC_8, KC_9,    KC_DEL,
+        KC_CAPS, _______, KC_LEFT, KC_DOWN, KC_RGHT, _______, _______, TD(TD_M_D), KC_4, KC_5, KC_6,    KC_ENT,
+        KC_LSFT, KC_MINS, KC_EQL,  KC_LBRC, KC_RBRC, KC_BSLS, _______, TD(TD_P_M), KC_1, KC_2, KC_3,    KC_RSFT,
+        _______, KC_LGUI, KC_LALT,                   _______,                            KC_0, KC_PDOT, XXXXXXX
     ),
 
-    [_LAYER2] = LAYOUT_DEFAULT(
+    [_LAYER2] = LAYOUT_625U_SPACE(
         KC_F1,   KC_F2,   KC_F3,   KC_F4,   KC_F5,   KC_6,    KC_F7,   KC_F8,   KC_F9,   KC_F10,  KC_F11,  KC_F12,
         XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, KC_VOLU,
-        XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, RESET,   XXXXXXX, XXXXXXX, KC_RCTL, KC_RALT, KC_DEL,  KC_VOLD,
-        XXXXXXX, _______, XXXXXXX, XXXXXXX,                   XXXXXXX,                   XXXXXXX, XXXXXXX, XXXXXXX
-    ),
-
-    [_LAYER3] = LAYOUT_DEFAULT(
-        KC_TILD, KC_EXLM, KC_AT,   KC_HASH, KC_DLR,  KC_PERC, KC_CIRC, KC_AMPR, KC_ASTR, KC_LPRN, KC_RPRN, KC_UNDS,
-        XXXXXXX, KC_P1,   KC_P2,   KC_P3,   KC_P4,   KC_P5,   KC_P6,   KC_P7,   KC_P8,   KC_P9,   KC_P0,   _______,
-        _______, KC_MINS, KC_EQL,  KC_LBRC, KC_RBRC, KC_BSLS, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, _______,
-        _______, XXXXXXX, XXXXXXX, XXXXXXX,                   _______,                   XXXXXXX, XXXXXXX, XXXXXXX
+        XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, RESET,   XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, XXXXXXX, KC_VOLD,
+        _______, XXXXXXX, XXXXXXX,                   XXXXXXX,                            KC_DEL,  KC_RALT, KC_RCTL
     ),
 };
+
 
 // Rotary Encoder Functions
 void encoder_update_user(uint8_t index, bool clockwise) {
