+++ keyboards/mb44/mb44.h
@@ -30,20 +30,8 @@
 #define XXX KC_NO
 
 
-#define LAYOUT_DEFAULT( \
-    K00, K01, K02, K03, K04, K05, K06, K07, K08, K09, K0A, K0B, \
-    K10, K11, K12, K13, K14, K15, K16, K17, K18, K19, K1A, K1B, \
-    K20, K21, K22, K23, K24, K25, K26, K27, K28, K29, K2A, K2B, \
-    K30, K31, K32, K33,           K36,           K39, K3A, K3B \
-) \
-{ \
-    { K00, K01, K02, K03, K04, K05, K06, K07, K08, K09, K0A, K0B }, \
-    { K10, K11, K12, K13, K14, K15, K16, K17, K18, K19, K1A, K1B }, \
-    { K20, K21, K22, K23, K24, K25, K26, K27, K28, K29, K2A, K2B }, \
-    { K30, K31, K32, K33, XXX, XXX, K36, XXX, XXX, K39, K3A, K3B } \
-}
 
-#define LAYOUT_625U_SPACE( \
+#define LAYOUT_DEFAULT( \
     K00, K01, K02, K03, K04, K05, K06, K07, K08, K09, K0A, K0B, \
     K10, K11, K12, K13, K14, K15, K16, K17, K18, K19, K1A, K1B, \
     K20, K21, K22, K23, K24, K25, K26, K27, K28, K29, K2A, K2B, \
@@ -80,4 +68,17 @@
     { K10, K11, K12, K13, K14, K15, K16, K17, K18, K19, K1A, K1B }, \
     { K20, K21, K22, K23, K24, K25, K26, K27, K28, K29, K2A, K2B }, \
     { K30, K31, K32, K33, XXX, K35, K36, XXX, XXX, K39, K3A, K3B } \
+}
+
+#define LAYOUT_3U_SPACE( \
+    K00, K01, K02, K03, K04, K05, K06, K07, K08, K09, K0A, K0B, \
+    K10, K11, K12, K13, K14, K15, K16, K17, K18, K19, K1A, K1B, \
+    K20, K21, K22, K23, K24, K25, K26, K27, K28, K29, K2A, K2B, \
+    K30, K31, K32, K33,           K36,           K39, K3A, K3B \
+) \
+{ \
+    { K00, K01, K02, K03, K04, K05, K06, K07, K08, K09, K0A, K0B }, \
+    { K10, K11, K12, K13, K14, K15, K16, K17, K18, K19, K1A, K1B }, \
+    { K20, K21, K22, K23, K24, K25, K26, K27, K28, K29, K2A, K2B }, \
+    { K30, K31, K32, K33, XXX, XXX, K36, XXX, XXX, K39, K3A, K3B } \
 }
